agregue validaciones para buscar por id, y eliminar, tambienselecciones los atributos que trae el get y los ordene

+++ src/controllers/blogController.js
@@ -1,16 +1,17 @@
 const router = require('express').Router()
-
-const {
-    Blog
-} = require("../../db")
+const {Blog} = require("../../db")
 
 router.get('/', async (req, res) => {
-    const blog = await Blog.findAll()
-    res.json(blog)
+    const blog = await Blog.findAll({attributes: ['id','titulo','imagen','categoria','fecha'],sort: '"fecha", descending'})
+    .then(blog=>res.json(blog))
 }) 
 router.get('/:id', async (req, res) => {
     const blog = await Blog.findByPk(req.params.id)
+    if(blog){
     res.json(blog)
+    }else{
+        res.json({succes: "id no encontrado"})
+    }
 }) 
 router.post('/', async (req, res) => {
    const blog = await Blog.create(req.body)
@@ -30,10 +31,16 @@ router.put('/:id', async (req, res) => {
  })
 
  router.delete('/:id', async (req, res) => {
-    await Blog.destroy({
+    const blog = await Blog.findByPk(req.params.id)
+    if(blog){ await Blog.destroy({
         where: { id: req.params.id}
     })
-    res.json({succes: "se ha eliminado"})
+    res.json({succes: "se ha eliminado"}) 
+}else{
+    res.json({succes: "id no encontrado"})
+
+}
+    
  })
 
 module.exports = router;
